Add the product configuring script
This python script is able to configure a wso2 product for a given DB. In other words, it modifyes given data sources in the product and integration pom files and compress the product. Thereafter this compressed product could be used to run integration tests.  The config.yml file is used to maintain configuration details of the product. based on the data in the configuration file, python script does modifications to the product.