Fix windows OS related issues
Regular DOS paths are limited to MAX_PATH (260) characters, including the string's terminating NUL character.  Therefore if we try to access file path which has more than 260 characters for its path, windows throw IOError. In order to exceed this limit, we can use extended-length paths. Hence in order to fix this issue checked the OS and, if OS is windows then modify the path to extended-length path.